% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobMulOutliers.R
\name{RobMulOutliers}
\alias{RobMulOutliers}
\title{Find robust multivariate outliers}
\usage{
RobMulOutliers(dat, dat_proportion, reject_alpha)
}
\arguments{
\item{dat}{Data frame of numeric (or coerced to numeric) variables}

\item{dat_proportion}{The proportion of data used for robust covariance estimation. Inversely related to number of shifted outliers.}

\item{reject_alpha}{The chi-squared alpha threshold for identifying outliers (a higher alpha leads to more shifted outliers).}
}
\value{
A list containing:\describe{
   \item{\code{excluded}}{Logical vector indicating whether each index is an outlier.}
   \item{\code{robCov}}{Robust estimate of covariance; output of \code{\link[MASS]{cov.mcd}}.}
   \item{\code{mDist}}{Numeric vector; robust Mahalanobis distance of each index from the multivariate center.}
   \item{\code{cutoff}}{Numeric scalar; given the \code{reject_alpha}, this is the Mahalanobis distance beyond which an index is considered to be an outlier.}
}
}
\description{
Implements the multivariate covariance estimation and outlier rejection 
approach advocated by Leys et al. (2018; \code{doi.org/10.1016/j.jesp.2017.09.011}). 
Used by \code{\link{RMW}} to manage outliers through robust multivariate Winsorization.
}
\examples{
iris_noOutliers <- RobMulOutliers(iris[,1:4],.9,.01)
}
