% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUW.R
\name{RUW}
\alias{RUW}
\title{Robust Univariate Winsorization}
\usage{
RUW(x, reject_alpha = pnorm(-3) * 2, method = "simple")
}
\arguments{
\item{x}{Numeric vector}

\item{reject_alpha}{The Gaussian alpha threshold for identifying outliers (a higher alpha leads to more shifted outliers). Defaults to an alpha corresponding to the density in both tails of a Gaussian cutoff of 3 standard deviations.}
}
\value{
Vector the same length as \code{x}, with the outliers of \code{x} Winsorized.
}
\description{
Winsorization involves the setting of some threshold for probable outliers 
(e.g., top 2 percent of cases or cases more than 4 MAD from the median) and replacing them
with values that correspond to that threshold (e.g., replacing a value 6 MAD from the median
with one 4 MAD from the median). This function implements a robust version of this
approach, following an asymmetric adaptation of the median-MAD approach advocated by Leys et al. 
(\code{doi.org/10.1016/j.jesp.2013.03.013}).
}
\details{
With \code{method=='simple'}: Given a numeric vector, find the median, the median deviation of the lower 50% of cases,
and the median deviation of the upper 50% of cases, and scale each asymmetric median
deviation to be equivalent to SD scale (the default of \code{\link[stats]{mad}}). 
Using some rejection alpha (i.e., tail density), identify
the number of Gaussian standard deviations \code{qThresh} corresponding to that alpha. 
For all values above \code{median + qThresh*upperMAD}, replace them 
with \code{median + qThresh*upperMAD}. For all values 
below \code{median - qThresh*lowerMAD}, replace them 
with \code{median - qThresh*lowerMAD}.

With \code{method=='distr'}: \emph{Not yet implemented}
}
\examples{
sepalWidthWins <- RUW(iris$Sepal.Width, reject_alpha = .01)
1-mean(sepalWidthWins == iris$Sepal.Width) # less than 2 percent of values Winsorized

}
